+++ angular-frontend/src/app/omnai-datasource/csv-file-import/csv-file-import.service.ts
@@ -22,6 +22,8 @@ export class CsvFileImportService implements DataSource {
   connect() {
     this.dialog.open(CsvFileSelectModalComponent)
   }
+  disconnect(): void { }
+  clearData(): void { }
   private readonly $data = signal<Record<string, DataFormat[]>>({});
   readonly data = this.$data.asReadonly();