+++ angular-frontend/src/app/source-selection/data-source-selection.service.ts
@@ -12,6 +12,8 @@ export interface DataPoint {
 /** Your expected DataSource interface */
 export interface DataSource {
     connect(): unknown;
+    disconnect(): void;
+    clearData(): void;
     data: Signal<Record<string, DataFormat[]>>
 }
 
@@ -37,6 +39,8 @@ export class DataSourceSelectionService {
             name: 'OmnAIScope',
             description: 'Live data from connected OmnAIScope devices',
             connect: this.liveDataService.connect.bind(this.liveDataService),
+            disconnect: this.liveDataService.disconnect.bind(this.liveDataService),
+            clearData: this.liveDataService.clearData.bind(this.liveDataService),
             data: this.liveDataService.data
         },
         {
@@ -44,14 +48,18 @@ export class DataSourceSelectionService {
             name: 'Random Dummy Data',
             description: 'Random generated data points',
             connect: this.dummyDataService.connect.bind(this.dummyDataService),
+            disconnect: this.dummyDataService.disconnect.bind(this.dummyDataService),
+            clearData: this.dummyDataService.clearData.bind(this.dummyDataService),
             data: this.dummyDataService.data
         },
         {
-          id: 'csv-file',
-          name: 'CSV Data',
-          description: 'Import a CSV file',
-          connect: this.csvDataService.connect.bind(this.csvDataService),
-          data: this.csvDataService.data
+            id: 'csv-file',
+            name: 'CSV Data',
+            description: 'Import a CSV file',
+            connect: this.csvDataService.connect.bind(this.csvDataService),
+            disconnect: this.csvDataService.disconnect.bind(this.csvDataService),
+            clearData: this.csvDataService.clearData.bind(this.csvDataService),
+            data: this.csvDataService.data
         }
     ]);
     readonly availableSources = this._availableSources.asReadonly();