+++ angular-frontend/src/app/source-selection/start-data-button.component.ts
@@ -11,16 +11,27 @@ import { AdvancedModeService } from '../advanced-mode/advanced-mode.service';
     standalone: true,
     imports: [MatDialogModule, MatIconModule],
     template: `
-    <button mat-icon-button (click)="openModal()" aria-label="Start Data">
-      <mat-icon>play_arrow</mat-icon>
-    </button>
-  `
+        <button mat-icon-button (click)="toggleStartButton()" aria-label="Start Data" id="start-button">
+        <mat-icon>{{ measurementIsStarted ? 'stop' : 'play_arrow' }}</mat-icon>
+        </button>
+    `,
+    styles: `button { display: flex; padding: .3em }`,
 })
 export class StartDataButtonComponent {
     private readonly dialog = inject(MatDialog);
     private readonly datasource = inject(DataSourceSelectionService);
     private readonly advancedMode = inject(AdvancedModeService);
+
+    protected measurementIsStarted: boolean = false;
+
+    clearAllData(): void {
+        this.datasource.availableSources().forEach( (source) => {
+            source.clearData();
+        });
+    }
+
     openModal(): void {
+        this.clearAllData();
         if (this.advancedMode.enabled()) {
             const dialogRef = this.dialog.open(SourceSelectModalComponent, {
                 width: '60vw'
@@ -39,5 +50,14 @@ export class StartDataButtonComponent {
                 OmnAIScope.connect();
             }
         }
+    }
+
+    stopMeasurement(): void {
+        this.datasource.currentSource()?.disconnect();
+    }
+
+    toggleStartButton(): void {
+        this.measurementIsStarted ? this.stopMeasurement() : this.openModal();
+        this.measurementIsStarted = !this.measurementIsStarted;
     }
 }