+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Formatting needs to be similar to the dataformat of the old OmnAIView data expor
 - Add darkmode via button toggle (#121)
 - Add workshop/advanced mode for car repair shops (#118)
 - Add CI for Release (#123)
+- Add stop and delete functionality to start button (#126)
 
 
 ### Changed